@@ -1,6 +1,6 @@
 {
-    "name": "mikecao/flight",
-    "description": "Flight is a fast, simple, extensible framework for PHP. Flight enables you to quickly and easily build RESTful web applications.",
+    "name": "flightphp/core",
+    "description": "Flight is a fast, simple, extensible framework for PHP. Flight enables you to quickly and easily build RESTful web applications. This is the maintained fork of mikecao/flight",
     "homepage": "http://flightphp.com",
     "license": "MIT",
     "authors": [
@@ -9,15 +9,72 @@
             "email": "mike@mikecao.com",
             "homepage": "http://www.mikecao.com/",
             "role": "Original Developer"
+        },
+        {
+            "name": "Franyer Sánchez",
+            "email": "franyeradriansanchez@gmail.com",
+            "homepage": "https://faslatam.000webhostapp.com",
+            "role": "Maintainer"
+        },
+        {
+            "name": "n0nag0n",
+            "email": "n0nag0n@sky-9.com",
+            "role": "Maintainer"
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": "^7.4|^8.0|^8.1|^8.2|^8.3",
+        "ext-json": "*"
     },
     "autoload": {
-        "files": [ "flight/autoload.php", "flight/Flight.php" ]
+        "files": [
+            "flight/autoload.php"
+        ]
+    },
+    "autoload-dev": {
+        "classmap": [
+            "tests/classes/User.php",
+            "tests/classes/Hello.php",
+            "tests/classes/Factory.php",
+            "tests/classes/TesterClass.php"
+        ]
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.6"
+        "ext-pdo_sqlite": "*",
+        "league/container": "^4.2",
+        "level-2/dice": "^4.0",
+        "phpstan/extension-installer": "^1.3",
+        "phpstan/phpstan": "^1.10",
+        "phpunit/phpunit": "^9.5",
+        "rregeer/phpunit-coverage-check": "^0.3.1",
+        "squizlabs/php_codesniffer": "^3.8"
+    },
+    "config": {
+        "allow-plugins": {
+            "phpstan/extension-installer": true
+        },
+        "process-timeout": 0,
+        "sort-packages": true
+    },
+    "scripts": {
+        "test": "phpunit",
+        "test-coverage": "rm -f clover.xml && XDEBUG_MODE=coverage vendor/bin/phpunit --coverage-html=coverage --coverage-clover=clover.xml && vendor/bin/coverage-check clover.xml 100",
+        "test-server": "echo \"Running Test Server\" && php -S localhost:8000 -t tests/server/",
+        "test-server-v2": "echo \"Running Test Server\" && php -S localhost:8000 -t tests/server-v2/",
+        "test-coverage:win": "del clover.xml && phpunit --coverage-html=coverage --coverage-clover=clover.xml && coverage-check clover.xml 100",
+        "lint": "phpstan --no-progress -cphpstan.neon",
+        "beautify": "phpcbf --standard=phpcs.xml",
+        "phpcs": "phpcs --standard=phpcs.xml -n",
+        "post-install-cmd": [
+            "php -r \"if (!file_exists('phpcs.xml')) copy('phpcs.xml.dist', 'phpcs.xml');\""
+        ]
+    },
+    "suggest": {
+        "latte/latte": "Latte template engine",
+        "tracy/tracy": "Tracy debugger",
+        "phpstan/phpstan": "PHP Static Analyzer"
+    },
+    "replace": {
+        "mikecao/flight": "2.0.2"
     }
 }
