@@ -1,77 +1,74 @@
 <?php
-/**
- * Flight: An extensible micro-framework.
- *
- * @copyright   Copyright (c) 2013, Mike Cao <mike@mikecao.com>
- * @license     MIT, http://flightphp.com/license
- */
 
-require_once 'vendor/autoload.php';
-require_once __DIR__.'/../flight/autoload.php';
+declare(strict_types=1);
 
-class RedirectTest extends PHPUnit_Framework_TestCase
+namespace tests;
+
+use flight\Engine;
+use PHPUnit\Framework\TestCase;
+
+class RedirectTest extends TestCase
 {
-    /**
-     * @var \flight\Engine
-     */
-    private $app;
+    private Engine $app;
 
-    function getBaseUrl($base, $url){
-        if ($base != '/' && strpos($url, '://') === false) {
-            $url = preg_replace('#/+#', '/', $base.'/'.$url);
+    protected function setUp(): void
+    {
+        $_SERVER['SCRIPT_NAME'] = '/subdir/index.php';
+
+        $this->app = new Engine();
+        $this->app->set('flight.base_url', '/testdir');
+    }
+
+    public function getBaseUrl($base, $url)
+    {
+        if ($base !== '/' && strpos($url, '://') === false) {
+            $url = preg_replace('#/+#', '/', $base . '/' . $url);
         }
 
         return $url;
     }
 
-    function setUp() {
-        $_SERVER['SCRIPT_NAME'] = '/subdir/index.php';
-
-        $this->app = new \flight\Engine();
-        $this->app->set('flight.base_url', '/testdir');
-    }
-
     // The base should be the subdirectory
-    function testBase(){
+    public function testBase()
+    {
         $base = $this->app->request()->base;
 
-        $this->assertEquals('/subdir', $base);
+        self::assertEquals('/subdir', $base);
     }
 
     // Absolute URLs should include the base
-    function testAbsoluteUrl(){
+    public function testAbsoluteUrl()
+    {
         $url = '/login';
         $base = $this->app->request()->base;
 
-        $this->assertEquals('/subdir/login', $this->getBaseUrl($base, $url));
+        self::assertEquals('/subdir/login', $this->getBaseUrl($base, $url));
     }
 
     // Relative URLs should include the base
-    function testRelativeUrl(){
+    public function testRelativeUrl()
+    {
         $url = 'login';
         $base = $this->app->request()->base;
 
-        $this->assertEquals('/subdir/login', $this->getBaseUrl($base, $url));
+        self::assertEquals('/subdir/login', $this->getBaseUrl($base, $url));
     }
 
     // External URLs should ignore the base
-    function testHttpUrl(){
+    public function testHttpUrl()
+    {
         $url = 'http://www.yahoo.com';
         $base = $this->app->request()->base;
 
-        $this->assertEquals('http://www.yahoo.com', $this->getBaseUrl($base, $url));
+        self::assertEquals('http://www.yahoo.com', $this->getBaseUrl($base, $url));
     }
 
     // Configuration should override derived value
-    function testBaseOverride(){
+    public function testBaseOverride()
+    {
         $url = 'login';
-        if ($this->app->get('flight.base_url') !== null) {
-            $base = $this->app->get('flight.base_url');
-        }
-        else {
-            $base = $this->app->request()->base;
-        }
+        $base = $this->app->get('flight.base_url') ?? $this->app->request()->base;
 
-        $this->assertEquals('/testdir/login', $this->getBaseUrl($base, $url));
+        self::assertEquals('/testdir/login', $this->getBaseUrl($base, $url));
     }
 }
