@@ -1,87 +1,88 @@
 <?php
-/**
- * Flight: An extensible micro-framework.
- *
- * @copyright   Copyright (c) 2012, Mike Cao <mike@mikecao.com>
- * @license     MIT, http://flightphp.com/license
- */
 
-require_once 'vendor/autoload.php';
-require_once __DIR__.'/../flight/autoload.php';
-require_once __DIR__.'/classes/User.php';
+declare(strict_types=1);
 
-class RegisterTest extends PHPUnit_Framework_TestCase
+namespace tests;
+
+use flight\Engine;
+use tests\classes\User;
+use PHPUnit\Framework\TestCase;
+
+class RegisterTest extends TestCase
 {
-    /**
-     * @var \flight\Engine
-     */
-    private $app;
+    private Engine $app;
 
-    function setUp() {
-        $this->app = new \flight\Engine();
+    protected function setUp(): void
+    {
+        $this->app = new Engine();
     }
 
     // Register a class
-    function testRegister(){
-        $this->app->register('reg1', 'User');
+    public function testRegister()
+    {
+        $this->app->register('reg1', User::class);
 
         $user = $this->app->reg1();
 
-        $this->assertTrue(is_object($user));
-        $this->assertEquals('User', get_class($user));
-        $this->assertEquals('', $user->name);
+        self::assertIsObject($user);
+        self::assertInstanceOf(User::class, $user);
+        self::assertEquals('', $user->name);
     }
 
     // Register a class with constructor parameters
-    function testRegisterWithConstructor(){
-        $this->app->register('reg2', 'User', array('Bob'));
+    public function testRegisterWithConstructor()
+    {
+        $this->app->register('reg2', User::class, ['Bob']);
 
         $user = $this->app->reg2();
 
-        $this->assertTrue(is_object($user));
-        $this->assertEquals('User', get_class($user));
-        $this->assertEquals('Bob', $user->name);
+        self::assertIsObject($user);
+        self::assertInstanceOf(User::class, $user);
+        self::assertEquals('Bob', $user->name);
     }
 
     // Register a class with initialization
-    function testRegisterWithInitialization(){
-        $this->app->register('reg3', 'User', array('Bob'), function($user){
+    public function testRegisterWithInitialization()
+    {
+        $this->app->register('reg3', User::class, ['Bob'], function ($user) {
             $user->name = 'Fred';
         });
 
         $user = $this->app->reg3();
 
-        $this->assertTrue(is_object($user));
-        $this->assertEquals('User', get_class($user));
-        $this->assertEquals('Fred', $user->name);
+        self::assertIsObject($user);
+        self::assertInstanceOf(User::class, $user);
+        self::assertEquals('Fred', $user->name);
     }
 
     // Get a non-shared instance of a class
-    function testSharedInstance() {
-        $this->app->register('reg4', 'User');
+    public function testSharedInstance()
+    {
+        $this->app->register('reg4', User::class);
 
         $user1 = $this->app->reg4();
         $user2 = $this->app->reg4();
         $user3 = $this->app->reg4(false);
 
-        $this->assertTrue($user1 === $user2);
-        $this->assertTrue($user1 !== $user3);
+        self::assertSame($user1, $user2);
+        self::assertNotSame($user1, $user3);
     }
 
     // Map method takes precedence over register
-    function testMapOverridesRegister(){
-        $this->app->register('reg5', 'User');
+    public function testMapOverridesRegister()
+    {
+        $this->app->register('reg5', User::class);
 
         $user = $this->app->reg5();
 
-        $this->assertTrue(is_object($user));
+        self::assertIsObject($user);
 
-        $this->app->map('reg5', function(){
+        $this->app->map('reg5', function () {
             return 123;
         });
 
         $user = $this->app->reg5();
 
-        $this->assertEquals(123, $user);
+        self::assertEquals(123, $user);
     }
 }
