@@ -1,9 +1,12 @@
 <?php
-require 'flight/Flight.php';
 
-Flight::route('/', function(){
+declare(strict_types=1);
+
+require 'flight/Flight.php';
+// require 'flight/autoload.php';
+
+Flight::route('/', function () {
     echo 'hello world!';
 });
 
 Flight::start();
-?>
