@@ -1,10 +1,18 @@
 <?php
-class Hello {
-    public function sayHi() {
+
+declare(strict_types=1);
+
+namespace tests\classes;
+
+class Hello
+{
+    public function sayHi(): string
+    {
         return 'hello';
     }
 
-    public static function sayBye() {
+    public static function sayBye(): string
+    {
         return 'goodbye';
     }
 }
