@@ -1,8 +1,15 @@
 <?php
-class User {
-    public $name;
 
-    public function User($name = ''){
+declare(strict_types=1);
+
+namespace tests\classes;
+
+class User
+{
+    public string $name;
+
+    public function __construct(string $name = '')
+    {
         $this->name = $name;
     }
 }
